make spinner reusable
